Modify Authorization and import tokens in the Api's

+++ language-center-app/src/main/java/Geeks/languagecenterapp/Controller/CourseController.java
@@ -5,15 +5,19 @@ import Geeks.languagecenterapp.DTO.Request.CourseRequest;
 import Geeks.languagecenterapp.DTO.Request.DayCourseRequest;
 import Geeks.languagecenterapp.DTO.Request.EnrollRequest;
 import Geeks.languagecenterapp.DTO.Response.CourseResponse;
+import Geeks.languagecenterapp.Model.Enum.UserAccountEnum;
 import Geeks.languagecenterapp.Model.UserEntity;
 import Geeks.languagecenterapp.Service.CourseService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/course")
@@ -22,32 +26,68 @@ public class CourseController {
     private CourseService courseService;
     //Create Course
     @PostMapping("/add")
-    public ResponseEntity<?> addCourse( @ModelAttribute CourseRequest body) throws JsonProcessingException {
+    public ResponseEntity<?> addCourse( @AuthenticationPrincipal UserEntity user ,@ModelAttribute CourseRequest body) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.add(body);
     }
     //update Course
     @PostMapping("/update/{id}")
-    public ResponseEntity<Object>updateCourse(@PathVariable("id") int id , @ModelAttribute CourseRequest body)throws JsonProcessingException{
+    public ResponseEntity<Object>updateCourse(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id , @ModelAttribute CourseRequest body)throws JsonProcessingException{
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.update(body, id);
     }
     //delete Course
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Object> deleteCourse(@PathVariable("id") int id ) throws JsonProcessingException {
+    public ResponseEntity<Object> deleteCourse(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id ) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.delete(id);
     }
     //Add Time and Day for A course
     @PostMapping("/add-day/{id}")
-    public ResponseEntity<Object> addTimeDay(@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> addTimeDay(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.addDay(body,id);
     }
     //update Time and Day for A course
     @PostMapping("/update-day/{id}")
-    public ResponseEntity<Object> updateTimeDay(@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> updateTimeDay(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.updateDay(body,id);
     }
     //delete Time and Day for A course
     @DeleteMapping("/delete-day/{id}")
-    public ResponseEntity<Object> deleteTimeDay(@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> deleteTimeDay(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id ,@ModelAttribute DayCourseRequest body) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.deleteDay(body,id);
     }
     //get All Courses
@@ -58,12 +98,24 @@ public class CourseController {
     }
     // Add Course to Favorite
     @PostMapping("/add-to-favorite/{id}")
-    public ResponseEntity<Object> addCourseToFavorite(@PathVariable("id") int id, @AuthenticationPrincipal UserEntity user) throws JsonProcessingException {
+    public ResponseEntity<Object> addCourseToFavorite(@PathVariable("id") int id, @AuthenticationPrincipal UserEntity user) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.USER){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.addToFavorite(id, user);
     }
     // Delete Course from Favorite
     @PostMapping("/remove-from-favorite/{id}")
-    public ResponseEntity<Object> deleteCourseFromFavorite(@PathVariable("id") int id, @AuthenticationPrincipal UserEntity user) throws JsonProcessingException {
+    public ResponseEntity<Object> deleteCourseFromFavorite(@PathVariable("id") int id, @AuthenticationPrincipal UserEntity user) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.USER){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.deleteFromFavorite(id, user);
     }
     // get Course Rate
@@ -74,12 +126,24 @@ public class CourseController {
     }
     // QR Attendance
     @PostMapping("/qr-attendance/{id}")
-    public ResponseEntity<Object> qrAttendance(@PathVariable("id") int id ,@ModelAttribute AttendanceRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> qrAttendance(@PathVariable("id") int id ,@ModelAttribute AttendanceRequest body,@AuthenticationPrincipal UserEntity user) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.TEACHER){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.qrAttendance(body,id);
     }
     //Manual Attendance
     @PostMapping("/manual-attendance/{id}")
-    public ResponseEntity<Object> manualAttendance(@PathVariable("id") int id ,@ModelAttribute EnrollRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> manualAttendance(@PathVariable("id") int id ,@ModelAttribute EnrollRequest body ,@AuthenticationPrincipal UserEntity user) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.TEACHER){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return courseService.manualAttendance(body,id);
     }
