+++ language-center-app/src/main/java/Geeks/languagecenterapp/Controller/QuizController.java
@@ -4,15 +4,19 @@ import Geeks.languagecenterapp.DTO.Request.QuestionQuizRequest;
 import Geeks.languagecenterapp.DTO.Request.QuestionRequest;
 import Geeks.languagecenterapp.DTO.Request.QuizRequest;
 import Geeks.languagecenterapp.DTO.Response.QuizResponse;
+import Geeks.languagecenterapp.Model.Enum.UserAccountEnum;
 import Geeks.languagecenterapp.Model.UserEntity;
 import Geeks.languagecenterapp.Service.QuizService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/quiz")
@@ -21,47 +25,101 @@ public class QuizController {
     private QuizService quizService;
     //add question
     @PostMapping("/add-question")
-    public ResponseEntity<?> addQuestion(@ModelAttribute QuestionRequest body)  {
+    public ResponseEntity<?> addQuestion(@AuthenticationPrincipal UserEntity user,@ModelAttribute QuestionRequest body)  {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.addQuestion(body);
     }
     //update question
     @PostMapping("/update-question/{id}")
-    public ResponseEntity<Object>updateQuestion(@PathVariable("id") int id , @ModelAttribute QuestionRequest body){
+    public ResponseEntity<Object>updateQuestion(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id , @ModelAttribute QuestionRequest body){
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.updateQuestion(body, id);
     }
     //delete question
     @DeleteMapping("/delete-question/{id}")
-    public ResponseEntity<Object> deleteQuestion(@PathVariable("id") int id ) {
+    public ResponseEntity<Object> deleteQuestion(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id ) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.deleteQuestion(id);
     }
     //add quiz
     @PostMapping("/add-quiz")
-    public ResponseEntity<?> addQuiz(@ModelAttribute QuizRequest body){
+    public ResponseEntity<?> addQuiz(@AuthenticationPrincipal UserEntity user,@ModelAttribute QuizRequest body){
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.addQuiz(body);
     }
     //update quiz
     @PostMapping("/update-quiz/{id}")
-    public ResponseEntity<Object>updateQuiz(@PathVariable("id") int id , @ModelAttribute QuizRequest body){
+    public ResponseEntity<Object>updateQuiz(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id , @ModelAttribute QuizRequest body){
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.updateQuiz(body, id);
     }
     //delete quiz
     @DeleteMapping("/delete-quiz/{id}")
-    public ResponseEntity<Object> deleteQuiz(@PathVariable("id") int id ){
+    public ResponseEntity<Object> deleteQuiz(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id ){
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.deleteQuiz(id);
     }
     //add question to quiz
     @PostMapping("/add-question-quiz/{id}")
-    public ResponseEntity<?> addQuizQuestion(@PathVariable("id") int id ,@ModelAttribute QuestionQuizRequest body)  {
+    public ResponseEntity<?> addQuizQuestion(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id ,@ModelAttribute QuestionQuizRequest body)  {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.addQuestionToQuiz(body,id);
     }
     //update question to quiz
     @PostMapping("/update-question-quiz/{id}")
-    public ResponseEntity<Object>updateQuizQuestion(@PathVariable("id") int id , @ModelAttribute QuestionQuizRequest body) {
+    public ResponseEntity<Object>updateQuizQuestion(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id , @ModelAttribute QuestionQuizRequest body) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.updateQuestionToQuiz(body, id);
     }
     //delete question from quiz
     @DeleteMapping("/delete-question-quiz/{id}")
-    public ResponseEntity<Object> deleteQuizQuestion(@PathVariable("id") int id , @ModelAttribute QuestionQuizRequest body)  {
+    public ResponseEntity<Object> deleteQuizQuestion(@AuthenticationPrincipal UserEntity user,@PathVariable("id") int id , @ModelAttribute QuestionQuizRequest body)  {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.deleteQuestionFromQuiz(body,id);
     }
     // Get All Quizzes with Questions
@@ -73,6 +131,12 @@ public class QuizController {
     //book a placement test
     @PostMapping("/take-quiz/{quizId}")
     public ResponseEntity<Object> bookPlacementTest(@AuthenticationPrincipal UserEntity user,@PathVariable("quizId") int id ) {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.USER){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return quizService.takeQuiz(user,id);
     }
 }
