+++ language-center-app/src/main/java/Geeks/languagecenterapp/Controller/ServiceController.java
@@ -2,14 +2,20 @@ package Geeks.languagecenterapp.Controller;
 
 import Geeks.languagecenterapp.DTO.Request.ServiceRequest;
 import Geeks.languagecenterapp.DTO.Response.ServiceWithCourseResponse;
+import Geeks.languagecenterapp.Model.Enum.UserAccountEnum;
 import Geeks.languagecenterapp.Model.ServiceEntity;
+import Geeks.languagecenterapp.Model.UserEntity;
 import Geeks.languagecenterapp.Service.ServiceService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/service")
@@ -19,17 +25,35 @@ public class ServiceController {
 
     //Create Service
     @PostMapping("/add")
-    public ResponseEntity<Object> addService(@ModelAttribute ServiceRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> addService(@AuthenticationPrincipal UserEntity user , @ModelAttribute ServiceRequest body) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return serviceService.add(body);
     }
     //update Service
     @PostMapping("/update/{id}")
-    public ResponseEntity<Object> updateService(@PathVariable("id") int id, @ModelAttribute ServiceRequest body) throws JsonProcessingException {
+    public ResponseEntity<Object> updateService(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id, @ModelAttribute ServiceRequest body) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return serviceService.update(body, id);
     }
     //delete Service
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Object> deleteService(@PathVariable("id") int id) throws JsonProcessingException {
+    public ResponseEntity<Object> deleteService(@AuthenticationPrincipal UserEntity user ,@PathVariable("id") int id) throws JsonProcessingException {
+        Map<String, String> response = new HashMap<>();
+        if (user.getAccountType()!= UserAccountEnum.ADMIN){
+            // Create a response object with the success message
+            response.put("message","You are UnAuthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
         return serviceService.delete(id);
     }
     //get All Services
@@ -40,7 +64,7 @@ public class ServiceController {
 
     // Get All Services with courses
     @GetMapping("/get/all/with-courses")
-    public ResponseEntity<List<ServiceWithCourseResponse>> getAllServicesAndCourses() {
+    public ResponseEntity<?> getAllServicesAndCourses() {
         List<ServiceWithCourseResponse> services = serviceService.getAllWithCourses();
         return ResponseEntity.ok(services);
     }
